Fixed another 32/64 bit error, thanks to Ben

+++ src/errors.d
@@ -16,7 +16,7 @@ class OratrParseException : OratrBaseException {
 }
 
 class OratrInvalidArgumentException : OratrBaseException {
-	this( string type, int position, string file = __FILE__, size_t line = __LINE__ )
+	this( string type, size_t position, string file = __FILE__, size_t line = __LINE__ )
     {
         super( "Invalid argument type \""~type~"\" in position "~to!string(position), file, line );
     }
